updated network headers

+++ modules/ustk_grabber/include/visp3/ustk_grabber/usNetworkGrabber.h
@@ -144,7 +144,7 @@ public:
     int samplingFrequency; // in Hz
 
     //transducer settings
-    int transducerRadius;
+    double transducerRadius;
     double scanLinePitch;
     unsigned int scanLineNumber;
